AccSoft/accsoft/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
